requirements.txt
streamlit
pandas
requests
google-generativeai
python-dotenv